requirements/testing.txt
-r pip.txt

# For linting tests
pycodestyle==2.4.0

pytest==4.2.0
pytest-django==3.4.7
django-dynamic-fixture==2.0.0
